@@ -3,15 +3,16 @@
   "description": "Cloud Functions for Firebase",
   "scripts": {
     "test": "jest '.*spec.js' --watchAll; exit 0",
+    "reinstall": "rm package-lock.json && rm -rf node_modules && npm i",
     "lint": "eslint .",
     "serve": "firebase emulators:start --only functions",
     "shell": "firebase functions:shell",
     "start": "npm run shell",
-    "deploy": "firebase deploy --only functions",
+    "deploy": "node -v && firebase deploy --only functions",
     "logs": "firebase functions:log"
   },
   "engines": {
-    "node": "12"
+    "node": "14"
   },
   "main": "index.js",
   "dependencies": {
@@ -19,7 +20,8 @@
     "child-process-promise": "^2.2.1",
     "dotenv": "^8.2.0",
     "firebase-admin": "^9.2.0",
-    "firebase-functions": "^3.13.2"
+    "firebase-functions": "^3.13.2",
+    "axios": "^0.21.1"
   },
   "devDependencies": {
     "@firebase/testing": "^0.20.11",
