@@ -1,6 +1,6 @@
-export default function({ store, redirect }) {
-    // If the user is not authenticated
-    if (!store.getters.loggedIn) {
-      return redirect('/login?redirect=/dashboard')
-    }
+export default function({ store, route, redirect }) {
+  // If the user is not authenticated
+  if (!store.getters.loggedIn) {
+    return redirect(`/login?redirect=${route.path}`);
   }
+}
