@@ -1,6 +1,16 @@
 <template>
   <div
-    class="px-2 lg:px-12 py-2 max-w-6xl mx-auto lg:py-6 bg-gray-800 text-indigo-100 min-h-screen"
+    class="
+      px-2
+      lg:px-12
+      py-2
+      max-w-6xl
+      mx-auto
+      lg:py-6
+      bg-gray-800
+      text-indigo-100
+      min-h-screen
+    "
   >
     <card class="shadow pt-3 pb-6 px-12 bg-gray-900 bg-opacity-50 mx-auto">
       <h1 class="mb-3 py-2 text-xl font-bold text-center">
@@ -21,7 +31,7 @@
 import URLSlugInput from "@/components/URLSlugInput";
 import Notebook from "@/components/Notebook";
 import Card from "@/components/Card";
-import { getNbJsonFromUrl } from "~/services/notebook";
+import { getNbJsonFromUrl } from "@/services/notebook";
 
 export default {
   name: "url-preview",
