@@ -1,15 +1,6 @@
+import splitbee from "@splitbee/web";
 
-export default async ({ app }, inject) => {
-      const EventName = app.$fireModule.analytics.EventName;
-
-      const analyticsEvents = {
-        ...EventName,
-        VIEW_USER_POSTS: "view_user_posts",
-        VIEW_USER_HOME: "view_user_home",
-        VIEW_POST: "view_post",
-        CREATE_POST: "create_post",
-        EDIT_POST: "edit_post",
-      }
-
-      inject('analyticsEvents', analyticsEvents)
-    }
+export default async (_, inject) => {
+  splitbee.init();
+  inject("splitbee", splitbee);
+};
